Add bulk user creation and update functionality with validation
