+++ packages/rest-typings/src/v1/users/UsersUpdateParamsPOST.ts
@@ -28,7 +28,7 @@ export type UsersUpdateParamsPOST = {
 	confirmRelinquish?: boolean;
 };
 
-const UsersUpdateParamsPostSchema = {
+export const UsersUpdateParamsPostSchema = {
 	type: 'object',
 	properties: {
 		userId: {
@@ -121,41 +121,3 @@ const UsersUpdateParamsPostSchema = {
 };
 
 export const isUsersUpdateParamsPOST = ajv.compile<UsersUpdateParamsPOST>(UsersUpdateParamsPostSchema);
-
-// Ultatel: Added Type (will move to new file later)
-
-export type UserBulkUpdateParamsPOST = Omit<UsersUpdateParamsPOST, 'userId'> & {
-	username: string;
-	data: UsersUpdateParamsPOST['data'] & {
-		extension: string;
-		companyPrefix: string;
-		companyId: number;
-		userId: number;
-	};
-};
-
-const userUpdateBulkPostSchema = {
-	type: 'array',
-	items: {
-		type: 'object',
-		properties: {
-			username: {
-				type: 'string',
-			},
-			data: {
-				...UsersUpdateParamsPostSchema.properties.data,
-				properties: {
-					...UsersUpdateParamsPostSchema.properties.data.properties,
-					extension: { type: 'string', nullable: false },
-					companyPrefix: { type: 'string', nullable: false },
-					companyId: { type: 'number', nullable: false },
-					userId: { type: 'number', nullable: false },
-				},
-			},
-		},
-		required: ['username', 'data'],
-		additionalProperties: false,
-	},
-};
-
-export const isUsersBulkUpdateParamsPOST = ajv.compile<UserBulkUpdateParamsPOST[]>(userUpdateBulkPostSchema);
